+++ lib/Settings/Admin.php
@@ -1,17 +1,18 @@
 <?php
+
 namespace OCA\Gitlab\Settings;
 
+use OCA\Gitlab\AppInfo\Application;
 use OCP\AppFramework\Http\TemplateResponse;
 use OCP\AppFramework\Services\IInitialState;
 use OCP\IConfig;
-use OCP\Settings\ISettings;
 
-use OCA\Gitlab\AppInfo\Application;
+use OCP\Settings\ISettings;
 
 class Admin implements ISettings {
 
 	public function __construct(private IConfig $config,
-								private IInitialState $initialStateService) {
+		private IInitialState $initialStateService) {
 	}
 
 	/**